+++ spotbugsTestCases/src/java/androidAnnotations/ObjectForNonNullParam.java
@@ -1,0 +1,12 @@
+package androidAnnotations;
+
+import android.support.annotation.NonNull;
+
+public class ObjectForNonNullParam {
+    static void foo(@NonNull Object o) {
+    }
+
+    static void bar() {
+        foo(new Object());
+    }
+}